requirements.txt
Django == 1.11.1
Scrapy == 1.4.0
djangorestframework == 3.7.7
scrapy_djangoitem == 1.1.1